+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 This list is not intended to be all-encompassing - it will document major and breaking API
 changes with their rationale when appropriate:
 
-### v5.4.1.0 (uncut)
+### v5.4.1.0
 - **http4k-*** : Upgrade some dependency versions.
 - **http4k-client-helidon** : API changes from Helidon alpha to M1
 - **http4k-server-helidon** : API changes from Helidon alpha to M1